Moved image cacheing to the adaptor level.
